+++ integration/e2emimir/client.go
@@ -1000,11 +1000,10 @@ func (c *Client) GetGrafanaAlertmanagerConfig(ctx context.Context) (*alertmanage
 	return ugc, err
 }
 
-func (c *Client) SetGrafanaAlertmanagerConfig(ctx context.Context, id, createdAtTimestamp int64, cfg, hash string, isDefault bool) error {
+func (c *Client) SetGrafanaAlertmanagerConfig(ctx context.Context, createdAtTimestamp int64, cfg, hash string, isDefault bool) error {
 	u := c.alertmanagerClient.URL("/api/v1/grafana/config", nil)
 
 	data, err := json.Marshal(&alertmanager.UserGrafanaConfig{
-		ID:                        id,
 		GrafanaAlertmanagerConfig: cfg,
 		Hash:                      hash,
 		CreatedAt:                 createdAtTimestamp,